+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.30.a1] - 2026-03-26
+## [0.30.a1] - 2026-03-27
 ## Added
 
 - Transfer stitched image to WSI view (low-res preview remains)
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add better permission on one-directory standalone builds
 - Update `dependabot` vulnerabilities: `requests`
+- Cache mask boundaries when first created for ROIs that are sufficiently large
+- Set dataset gallery down-sampling default to `True`
 
 ## [0.29.0] - 2026-03-17